+++ sphaera/diffeq/_impl/adams.py
@@ -73,9 +73,9 @@ class VariableCoefficientAdamsBashforth(AdaptiveStepsizeODESolver):
         self.atol = atol if _is_iterable(atol) else [atol] * len(y0)
         self.implicit = implicit
         self.max_order = int(max(_MIN_ORDER, min(max_order, _MAX_ORDER)))
-        self.safety = _convert_to_tensor(safety, dtype=torch.float32, device=y0[0].device)
-        self.ifactor = _convert_to_tensor(ifactor, dtype=torch.float32, device=y0[0].device)
-        self.dfactor = _convert_to_tensor(dfactor, dtype=torch.float32, device=y0[0].device)
+        self.safety = _convert_to_tensor(safety, dtype=torch.float64, device=y0[0].device)
+        self.ifactor = _convert_to_tensor(ifactor, dtype=torch.float64, device=y0[0].device)
+        self.dfactor = _convert_to_tensor(dfactor, dtype=torch.float64, device=y0[0].device)
 
     def before_integrate(self, t):
         prev_f = collections.deque(maxlen=self.max_order + 1)
